gcry_mpi_scan() may decide that the actual size of the number passed to it to convert into an MPI is smaller that the 'size' argument passed to it along with the number, in which case we get a wrong key size during resume and the resume fails as a result.
To fix this, make the key generation code check if the size of the session key
and ivec is correct from the gcry_mpi_scan()/gcry_mpi_aprint() perspective.

+++ encrypt.h
@@ -44,8 +44,8 @@ struct RSA_data {
 };
 
 struct encrypted_key {
-	unsigned short	size;
-	unsigned char	data[KEY_DATA_SIZE];
+	size_t size;
+	unsigned char data[KEY_DATA_SIZE];
 };
 
 struct key_data {
